+++ game/content/ghrooms.py
@@ -118,14 +118,14 @@ class BarArea(OpenRoom):
         super().build(gb, archi)
 
         # Add a bar along the south and maybe along one side.
-        mydest = pygame.Rect(self.area.left, self.area.bottom - 1, self.area.width, 1)
+        mydest = pygame.Rect(self.area.left + 1, self.area.bottom - 1, self.area.width - 2, 1)
         gb.fill(mydest, wall=self.COUNTER_TYPE)
 
         if random.randint(1, 3) == 1:
-            mydest = pygame.Rect(self.area.left, self.area.top, 1, self.area.height)
+            mydest = pygame.Rect(self.area.left, self.area.top + 1, 1, self.area.height - 1)
             gb.fill(mydest, wall=self.COUNTER_TYPE)
         elif random.randint(1, 2) == 1:
-            mydest = pygame.Rect(self.area.right, self.area.top, 1, self.area.height)
+            mydest = pygame.Rect(self.area.right, self.area.top + 1, 1, self.area.height - 1)
             gb.fill(mydest, wall=self.COUNTER_TYPE)